requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-crispy-forms==1.13.0
django-import-export==2.7.1
django-inspect-model==0.7
django-progressive-web-app==0.1.1
et-xmlfile==1.1.0
MarkupPy==1.14
mssql-django==1.1.1
odfpy==1.4.1
openpyxl==3.0.9
pyodbc==4.0.32
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tablib==3.1.0
wfastcgi==3.0.0
xlrd==2.0.1
xlwt==1.3.0
